gr-analog: Fix PSD computation errors: cast noverlap to int

+++ gr-analog/examples/fmtest.py
@@ -142,7 +142,7 @@ def main():
         d = fm.snk_tx.data()[Ns:Ns + Ne]
         sp1_f = fig1.add_subplot(2, 1, 1)
 
-        X, freq = sp1_f.psd(d, NFFT=fftlen, noverlap=fftlen / 4, Fs=fs,
+        X, freq = sp1_f.psd(d, NFFT=fftlen, noverlap= int(fftlen / 4), Fs=fs,
                             window=lambda d: d * winfunc(fftlen),
                             visible=False)
         X_in = 10.0 * numpy.log10(abs(numpy.fft.fftshift(X)))
@@ -181,7 +181,7 @@ def main():
             d = fm.snks[i].data()[Ns:Ne]
 
             sp2_f = fig2.add_subplot(Nrows, Ncols, 1 + i)
-            X, freq = sp2_f.psd(d, NFFT=fftlen, noverlap=fftlen / 4, Fs=fs_o,
+            X, freq = sp2_f.psd(d, NFFT=fftlen, noverlap= int(fftlen / 4), Fs=fs_o,
                                 window=lambda d: d * winfunc(fftlen),
                                 visible=False)
             #X_o = 10.0*numpy.log10(abs(numpy.fft.fftshift(X)))
